requirements.txt
beautifulsoup4==4.12.3
groq==0.11.0
Markdown==3.7
openai==1.52.2
pandas==2.2.3
python-dotenv==1.0.1
reportlab==4.2.5
streamlit==1.39.0
